Fixes bug where equip commands are not parsed correctly
More specifically, parseEquipItem was parsing formIds as signed integers
using atoi(). When the number parsed surpasses 0x7fffffff, it will be
capped there. This will cause the formId lookup to fail and for the
equip command to not run.  The nature of this bug means it will only show
up when the thing being equipped is from an esp with a mod index >=128.

This commit fixes this issue by using strtoul().